Guitar Basics and LCD Tuner
This code is designed to be the basic structure of the guitar. Instead of piezo-electric sensors for inputs, this code just checks six buttons to see when a note has been played. Instead of ribbon potentiometers on the neck of the guitar it just check six knob potentiometers. This code also contains the basic code for interfacing the LCD tuning screen. Each string can be tuned up or down half steps according to button input.