Knock Sensor
This code was useful in getting the piezo inputs to work as was required.

This code comes directly from Arduino.cc. It is designed to read analog input of a piezo sensor. If the tap on the sensor is sufficient to pass the indicated threshold, the serial output will indicate it.